refactor(customers): restructure folders for greater cohesion
