+++ customers/src/main/java/io/github/joaoVitorLeal/marketsphere/customers/client/brasilapi/config/BrasilApiFeignConfig.java
@@ -1,6 +1,6 @@
-package io.github.joaoVitorLeal.marketsphere.customers.client.config;
+package io.github.joaoVitorLeal.marketsphere.customers.client.brasilapi.config;
 
-import io.github.joaoVitorLeal.marketsphere.customers.client.error.BrasilApiErrorDecoder;
+import io.github.joaoVitorLeal.marketsphere.customers.client.brasilapi.error.BrasilApiErrorDecoder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
