+++ customers/src/main/java/io/github/joaoVitorLeal/marketsphere/customers/client/brasilapi/representation/BrasilApiAddressRepresentation.java
@@ -1,10 +1,10 @@
-package io.github.joaoVitorLeal.marketsphere.customers.dto.brasilapi;
+package io.github.joaoVitorLeal.marketsphere.customers.client.brasilapi.representation;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record BrasilApiAddressDto(
+public record BrasilApiAddressRepresentation(
 
         @JsonAlias("cep")
         String postalCode,
